@@ -777,6 +777,80 @@ namespace UndertaleModLib
     {
         public override string Name => "BGND";
 
+        private bool checkedFor2024_14_1 = false;
+        private void CheckForGM2024_14_1(UndertaleReader reader)
+        {
+            checkedFor2024_14_1 = true;
+
+            if (!reader.undertaleData.IsVersionAtLeast(2024, 13) || reader.undertaleData.IsVersionAtLeast(2024, 14, 1))
+            {
+                return;
+            }
+
+            long returnTo = reader.Position;
+            long chunkStartPos = reader.AbsPosition;
+
+            // Go through each background, and check to see if it ends at the expected position. If not, this is probably 2024.14.1.
+            uint bgCount = reader.ReadUInt32();
+            for (int i = 0; i < bgCount; i++)
+            {
+                // Find background's start position, and calculate next background position (if available).
+                reader.Position = returnTo + 4 + (4 * i);
+                uint bgPtr = reader.ReadUInt32();
+                if (bgPtr == 0)
+                {
+                    // Removed asset
+                    continue;
+                }
+                uint nextBgPtr = 0;
+                int j = i;
+                while (nextBgPtr == 0 && (++j) < bgCount)
+                {
+                    // Try next pointer in list
+                    nextBgPtr = reader.ReadUInt32();
+                }
+
+                // Skip all the way to "GMS2ItemsPerTileCount" (at its pre-2024.14.1 location), which is what we actually care about.
+                reader.AbsPosition = bgPtr + (11 * 4);
+                uint itemsPerTileCount = reader.ReadUInt32();
+                uint tileCount = reader.ReadUInt32();
+
+                // Calculate the theoretical end position given the above info, and compare to the actual end position (with padding).
+                uint theoreticalEndPos = bgPtr + (15 * 4) + (itemsPerTileCount * tileCount * 4);
+                if (nextBgPtr == 0)
+                {
+                    // Align to 16 bytes, and compare against chunk end position
+                    if ((theoreticalEndPos % 16) != 0)
+                    {
+                        theoreticalEndPos += 16 - (theoreticalEndPos % 16);
+                    }
+                    uint chunkEndPos = (uint)chunkStartPos + Length;
+                    if (theoreticalEndPos != chunkEndPos)
+                    {
+                        // Probably 2024.14.1!
+                        reader.undertaleData.SetGMS2Version(2024, 14, 1);
+                        break;
+                    }
+                }
+                else
+                {
+                    // Align to 8 bytes, and compare against next background start position
+                    if ((theoreticalEndPos % 8) != 0)
+                    {
+                        theoreticalEndPos += 8 - (theoreticalEndPos % 8);
+                    }
+                    if (theoreticalEndPos != nextBgPtr)
+                    {
+                        // Probably 2024.14.1!
+                        reader.undertaleData.SetGMS2Version(2024, 14, 1);
+                        break;
+                    }
+                }
+            }
+
+            reader.Position = returnTo;
+        }
+
         internal override void SerializeChunk(UndertaleWriter writer)
         {
             Alignment = 8;
@@ -786,7 +860,22 @@ namespace UndertaleModLib
         internal override void UnserializeChunk(UndertaleReader reader)
         {
             Alignment = 8;
+
+            if (!checkedFor2024_14_1)
+            {
+                CheckForGM2024_14_1(reader);
+            }
+
             base.UnserializeChunk(reader);
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            checkedFor2024_14_1 = false;
+
+            CheckForGM2024_14_1(reader);
+
+            return base.UnserializeObjectCount(reader);
         }
     }
 
@@ -1090,7 +1179,7 @@ namespace UndertaleModLib
         {
             checkedFor2024_14 = true;
 
-            if (!reader.undertaleData.IsVersionAtLeast(2024, 13))// || reader.undertaleData.IsVersionAtLeast(2024, 14))
+            if (!reader.undertaleData.IsVersionAtLeast(2024, 13) || reader.undertaleData.IsVersionAtLeast(2024, 14))
             {
                 return;
             }
@@ -1598,6 +1687,7 @@ namespace UndertaleModLib
                     {
                         reader.SubmitWarning("Missing expected TPAG padding");
                         reader.Position--;
+                        break;
                     }
                 }
             }
