@@ -325,8 +325,10 @@ namespace UndertaleModTool
 
             if (Keyboard.Modifiers.HasFlag(ModifierKeys.Control))
             {
-                gridWidth /= 2;
-                gridHeight /= 2;
+                if (gridWidth > 1)
+                    gridWidth /= 2;
+                if (gridHeight > 1)
+                    gridHeight /= 2;
             }
             else if (Keyboard.Modifiers.HasFlag(ModifierKeys.Shift))
             {
@@ -509,7 +511,8 @@ namespace UndertaleModTool
                         ScaleX = other.ScaleX,
                         ScaleY = other.ScaleY,
                         Color = other.Color,
-                        Rotation = other.Rotation
+                        Rotation = other.Rotation,
+                        InstanceID = other.InstanceID
                     };
                     undoStack.Push(undoObj);
                     hotpointX = mousePos.X - other.X;
@@ -672,6 +675,7 @@ namespace UndertaleModTool
                 var newPartSysInst = new ParticleSystemInstance
                 {
                     Name = ParticleSystemInstance.GenerateRandomName(mainWindow.Data),
+                    InstanceID = ++mainWindow.Data.LastParticleSystemInstanceID,
                     ParticleSystem = partSysInst.ParticleSystem,
                     X = (int)pos.X,
                     Y = (int)pos.Y,
@@ -821,8 +825,10 @@ namespace UndertaleModTool
                 }
 
                 // Snap to grid
-                tgtX = ((tgtX + gridWidth / 2) / gridWidth) * gridWidth;
-                tgtY = ((tgtY + gridHeight / 2) / gridHeight) * gridHeight;
+                if (gridWidth > 0)
+                    tgtX = ((tgtX + gridWidth / 2) / gridWidth) * gridWidth;
+                if (gridHeight > 0)
+                    tgtY = ((tgtY + gridHeight / 2) / gridHeight) * gridHeight;
 
                 if (movingObj is GameObject gameObj)
                 {
