@@ -1,15 +1,15 @@
-﻿using Microsoft.Win32;
-using System;
+﻿using System;
+using System.ComponentModel;
 using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using ImageMagick;
+using Microsoft.Win32;
 using UndertaleModLib.Models;
 using UndertaleModLib.Util;
-using System.Windows.Controls;
-using System.Windows.Media;
-using System.Windows.Data;
 using UndertaleModTool.Windows;
-using ImageMagick;
-using System.Windows.Media.Imaging;
-using System.ComponentModel;
 
 namespace UndertaleModTool
 {
@@ -158,7 +158,18 @@ namespace UndertaleModTool
             {
                 using MagickImage image = TextureWorker.ReadBGRAImageFromFile(dlg.FileName);
                 UndertaleTexturePageItem item = DataContext as UndertaleTexturePageItem;
+
+                var previousFormat = item.TexturePage.TextureData.Image.Format;
+
                 item.ReplaceTexture(image);
+
+                var currentFormat = item.TexturePage.TextureData.Image.Format;
+
+                // If texture was DDS, warn user that texture has been converted to PNG
+                if (previousFormat == GMImage.ImageFormat.Dds && currentFormat == GMImage.ImageFormat.Png)
+                {
+                    mainWindow.ShowMessage($"{item.TexturePage} was converted into PNG format since we don't support converting images into DDS format. This might have performance issues in the game.");
+                }
 
                 // Refresh the image of "ItemDisplay"
                 if (ItemDisplay.FindName("RenderAreaBorder") is not Border border)
